mcp/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "psmcp"
version = "0.85.1"
description = "Adobe Photoshop automation using MCP"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Mike Chambers", email = "mikechambers@gmail.com"}
]
dependencies = [
    "fonttools",
    "python-socketio",
    "mcp[cli]",
    "requests",
    "websocket-client>=1.8.0",
    "pillow>=11.2.1",
    "numpy>=2.2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "black",
    "isort",
    "mypy",
]

[tool.setuptools]
py-modules = ["fonts", "logger", "psmcp", "socket_client"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
